put RotateUserAgent and Generation of Cookies into one Module
