+++ smzdm/smzdm/items.py
@@ -1,16 +1,9 @@
 # -*- coding: utf-8 -*-
 
-# Define here the models for your scraped items
-#
-# See documentation in:
-# http://doc.scrapy.org/en/latest/topics/items.html
-
 import scrapy
 
 
 class SmzdmItem(scrapy.Item):
-    # define the fields for your item here like:
-    # name = scrapy.Field()
     title = scrapy.Field()
     price = scrapy.Field()
     brief = scrapy.Field()
