+++ smzdm/smzdm/pipelines.py
@@ -1,11 +1,5 @@
 # -*- coding: utf-8 -*-
 
-# Define your item pipelines here
-#
-# Don't forget to add your pipeline to the ITEM_PIPELINES setting
-# See: http://doc.scrapy.org/en/latest/topics/item-pipeline.html
-
-
 import pymongo
 
 from scrapy.conf import settings
